don't try to delete the app if deployment was not requested

+++ app_build_suite/build_steps/base_test_runner.py
@@ -402,6 +402,8 @@ class BaseTestRunner(BuildStep, ABC):
 
     # noinspection PyMethodMayBeStatic
     def _delete_app(self, config: argparse.Namespace, context: Context):
+        if get_config_value_by_cmd_line_option(config, BaseTestRunnersFilteringPipeline.key_config_option_deploy_app):
+            return
         app_obj = cast(AppCR, context[context_key_app_cr])
         app_obj.delete()
         app_config_file_path = get_config_value_by_cmd_line_option(
